remove shared folder
