+++ src/components/DiscoveryGrid.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { View, StyleSheet, TouchableOpacity } from 'react-native'
 import Grid from 'react-native-grid-component'
 
-import { IMAGE_URL } from '../../utils/helperFunctions'
+import { IMAGE_URL } from '../utils/helperFunctions'
 import { DiscoveryCell } from './DiscoveryCell'
-import { SECTIONS } from '../../utils/constants'
+import { SECTIONS } from '../utils/constants'
 
 const _renderItem = (data, i) => (
   <TouchableOpacity
