+++ src/components/Tagline.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import { View, StyleSheet, Text } from 'react-native'
-import { PublicationPrimaryColor } from '../../utils/branding'
-import { BODY_SERIF } from '../../utils/fonts'
+
+import { PublicationPrimaryColor } from '../utils/branding'
+import { BODY_SERIF } from '../utils/fonts'
 
 const styles = StyleSheet.create({
   tagline: {
