+++ src/components/index.js
@@ -10,6 +10,6 @@ export * from './PictureHeadline'
 export * from './SectionHeader'
 export * from './Tagline'
 export * from './VerticalArticleCard'
-export * from './SectionView'
 export * from './ActivityIndicator'
 export * from './PrimaryHorizontalArticleCell'
+export * from './SearchBar'
