+++ src/utils/helperFunctions.js
@@ -1,6 +1,4 @@
-import { NavigationHelpersContext } from '@react-navigation/native'
 import moment from 'moment'
-import React from 'react'
 
 export const IMAGE_URL = (attachment_uuid, extension) =>
   `https://snworksceo.imgix.net/dpn/${attachment_uuid}.sized-1000x1000.${extension}?w=1000`
